Move initial qcengine runner to test utils
